refactor(microwave): 🛠 add applyStrategy function
