+++ packages/microwave/src/lib/microwave.ts
@@ -1,7 +1,7 @@
 import { distinctUntilChanged, Observable, takeUntil } from 'rxjs';
+import { applyStrategy } from './apply-strategy';
 import { getEngine } from './core/engine';
-import { decorateComponent, IvyComponentType } from './decorator';
-import { getStrategyDevKit } from './devkit';
+import { IvyComponentType } from './shared/decorator';
 import { asapStrategy } from './strategies/asap';
 
 /**
@@ -9,11 +9,7 @@ import { asapStrategy } from './strategies/asap';
  */
 export function Microwave(strategy = asapStrategy) {
   return function MicrowaveDecorator<T>(componentType: IvyComponentType<T>) {
-    _bindComponentToEngine(componentType, {
-      onCreate(component) {
-        strategy(getStrategyDevKit(component));
-      },
-    });
+    applyStrategy(componentType, strategy);
   };
 }
 
@@ -29,28 +25,4 @@ export function watch<T, K extends keyof T = keyof T>(
     distinctUntilChanged(),
     takeUntil(destroyed$)
   );
-}
-
-export function _bindComponentToEngine<T>(
-  componentType: IvyComponentType<T>,
-  { onCreate }: { onCreate: (component: T) => void }
-) {
-  decorateComponent(componentType, {
-    onCreate(component, changeDetectionFns) {
-      getEngine(component).setChangeDetectionFns(changeDetectionFns);
-      onCreate(component);
-    },
-    onDestroy(component) {
-      getEngine(component).markDestroyed();
-    },
-    onPropertyDeclare(component, property, value) {
-      getEngine(component).setProperty(property, value);
-    },
-    onPropertyGet(component, property) {
-      return getEngine(component).getProperty(property);
-    },
-    onPropertySet(component, property, value) {
-      getEngine(component).setProperty(property, value);
-    },
-  });
 }
